Binary Function XXXX

The most basic algorithm is NAND, that is the atomic building block of all higher-level algorithms.
A second abstraction (or complexity) layer is composed of the binary functions (aka BiFun) of dimension 1x1 and 2x1.
These are NAND trees that take 1 or 2 input bits and return 1 output bit.
There are \(2^2 + 2^4 = 4 + 16 = 20\) distinct such functions.
These are the building block functions from which any more complex algorithm may be built.
There is one specialized R6 class for every binary function, in the form BiFun_xxxx where "xxxx" is the binary number representation of the algorithm's truth table.
Amongst these, some are particularly useless, especially:

BiFun_00: The constant 0.

BiFun_11: The constant 1.

BiFun_0000: The constant 0.

BiFun_1111: The constant 1.

There is a generic R6 class BiFun whose constructor takes a truthtable parameter of any supported type that builds its internal NAND tree on the fly. Finally, well-known logical gates have their dedicated alias classes:

BiFun_AND,

BiFun_NAND,

BiFun_NOT,

BiFun_OR,

BiFun_XNOR,

BiFun_XOR.

BiFun_NAND

AlgoTTxxxx

Format

An object of class R6ClassGenerator of length 24.

Examples

# Specialized BiFun_xxxx R6 classes:
BiFun_00$new();

#> Error in eval(expr, envir, enclos): object 'BiFun_00' not found

BiFun_10$new();

#> Error in eval(expr, envir, enclos): object 'BiFun_10' not found

BiFun_0000$new();

#> Error in eval(expr, envir, enclos): object 'BiFun_0000' not found

BiFun_1000$new();

#> Error in eval(expr, envir, enclos): object 'BiFun_1000' not found

BiFun_0100$new();

#> Error in eval(expr, envir, enclos): object 'BiFun_0100' not found

# etc.

# Generic BiFun R6 class:
BiFun$new(truthtable = "00");

#> Error in eval(expr, envir, enclos): object 'BiFun' not found

BiFun$new(truthtable = "10");

#> Error in eval(expr, envir, enclos): object 'BiFun' not found

BiFun$new(truthtable = "0000");

#> Error in eval(expr, envir, enclos): object 'BiFun' not found

BiFun$new(truthtable = "1000");

#> Error in eval(expr, envir, enclos): object 'BiFun' not found

BiFun$new(truthtable = "0100");

#> Error in eval(expr, envir, enclos): object 'BiFun' not found

# etc.

# Well-known logical gates R6 class:
BiFun_AND$new();

#> Error in eval(expr, envir, enclos): object 'BiFun_AND' not found

BiFun_XOR$new();

#> Error in eval(expr, envir, enclos): object 'BiFun_XOR' not found