do_compute_vertex_name

Computes the unique name of an igraph vertex used in algo_composite internals. The naming scheme goes like this: For input bits: For algo: For output bits: This function is mainly used for internal purposes but could be useful for advanced manipulations. This function is trivial but may prove useful in the future to enforce data quality validation rules.

do_compute_vertex_name(id, bit, ...)

Arguments

id
The node ID character)

bit
The bit (character)

...
For future usage

Value

N/A

Examples

print(do_compute_vertex_name("algo1", "i3"));

#> [1] "algo1.i3"

print(do_compute_vertex_name("algo2", paste0("i", 1:3)));

#> [1] "algo2.i1" "algo2.i2" "algo2.i3"