Notation

The following notation conventions are used:

Sets

\(\mathbb{B} \): The set of all modular binary numbers.

\(\mathbb{B}_{1} = {0,1} \): The set of all modular binary numbers of dimension 1.

\(\mathbb{B}_{2} = {00,10,01,11} \): The set of all modular binary numbers of dimension 2.

\(\mathbb{B}_{i} = {00...i,10...i,...} \): The set of all modular binary numbers of dimension i.

\(\mathbb{A}\): The set of all algorithms.

\(\mathbb{A}^{\mathbb{B}}\): The set of all algorithms that receive modular binary numbers as input and output modular binary numbers.

\(\mathbb{A}^{\mathbb{B}_{2},\mathbb{B}_{1}}\): The set of all algorithms that receive modular binary numbers of dimension 2 as input and output modular binary numbers of dimension 1.

\(\mathbb{A}^{\mathbb{B}_{i},\mathbb{B}_{j}}\): The set of all algorithms that receive modular binary numbers of dimension i as input and output modular binary numbers of dimension j.

\(\mathbb{T}^{\mathbb{A}}\): The set of all transformations that receive algorithms as inputs and output algorithms.

Modular Binary Number Constants

Modular binary numbers are usually expressed as a sequence of "0" and "1" bits with the left-most bit being the least significant bit. Examples:

\(000\)

\(110\)

\(0000\)

\(001100\)

If the context is ambiguous due to the presence of integers, modular binary numbers may be prefixed with a "b". Examples:

\(b000 \)

\(b110 \)

\(b0000 \)

\(b001100 \)

Variables

\(i,j,k \in \mathbb{N}^+\): Variables representing dimensions.

\(n \in \mathbb{B}\): Variables representing modular binary numbers.

\(a \in \mathbb{A}^{\mathbb{B}}\): Variables representing algorithms.

\(t \in \mathbb{T}^{\mathbb{A}x\mathbb{A}}\): Variables representing algorithms.

Bits of modular binary numbers

To reference bits of modular binary numbers, we use the dot notation.

\(n.2\): The second bit of n, starting from left.

\(n.i\): The ith bit of n, starting from left.

Input / output bits of algorithms

To reference input / output bits of algorithms, we use the dot notation.

\(a.i_{3}\): The 3rd input bit of a.

\(a.i_{i}\): The ith input bit of a.

\(a.o_{2}\): The 2nd output bit of a.

\(a.o_{i}\): The ith output bit of a.

Components of algorithm composites

To reference components of algorithm composites, we use the dot notation:

\(a_{composite}.a_{1}\): The first sub-algorithm of composite a.

\(a_{composite}.a_{i}\): The ith sub-algorithms of composite a.

\(a_{1}.o_{1} \rightarrow a_{2}.i_{1}\): The edge linking the output bit 1 of algo 1 to input bit 1 of algo 2.

Component predecessors and successors

To reference predecessors and successors of algorithm components, we use:

\(N^{-}(x)\): The immediate predecessors of x.

\(N^{+}(x)\): The immediate successors of x.

Dimension

\(dim(n)\): The dimension of n.

\(dim_{i}(a)\): The input dimension of a.

\(dim_{o}(a)\): The output dimension of b.