Transformation: truth tables to DAG

Transformation: truth tables to DAG

Examples

# Create a random truth table algorithm with input dimension 8 and output dimension 8:
truthtable <- algo_tt$new(8,8)$do_randomize_outputs();
# Display the upper part of the truth table in the console:
cat(substr(truthtable$get_prettystring(),1,200));

#> 00000000 > 10101000
#> 10000000 > 11001100
#> 01000000 > 10101011
#> 11000000 > 11011101
#> 00100000 > 01110100
#> 10100000 > 00011011
#> 01100000 > 01000001
#> 11100000 > 00010110
#> 00010000 > 01000100
#> 10010000 > 10111110

# Transform the truth table to a DAG algorithm:
dag <- atomize(truthtable);

#> INFO [2018-09-02 19:49:20] set_dim_i: increase input dimension by 8
#> INFO [2018-09-02 19:49:20] set_dim_o: increase output dimension by 8
#> INFO [2018-09-02 19:49:20] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:20] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:20] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:20] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:20] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:20] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:20] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:20] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:21] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:21] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1
#> INFO [2018-09-02 19:49:22] set_dim_i: increase input dimension by 2
#> INFO [2018-09-02 19:49:22] set_dim_o: increase output dimension by 1

#